refactor midas folder structure
