+++ midas/backbones/vit.py
@@ -5,8 +5,8 @@ import types
 import math
 import torch.nn.functional as F
 
-from .new_backbones.utils import (activations, forward_adapted_unflatten, get_activation, get_readout_oper,
-                                  make_backbone_default, Transpose)
+from .utils import (activations, forward_adapted_unflatten, get_activation, get_readout_oper,
+                    make_backbone_default, Transpose)
 
 
 def forward_vit(pretrained, x):
